# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/paymentItems": {
            "get": {
                "tags": [
                    "PaymentItem"
                ],
                "summary": "Get a listing of the PaymentItems.",
                "description": "Get all PaymentItems",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/PaymentItem"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "PaymentItem"
                ],
                "summary": "Store a newly created PaymentItem in storage",
                "description": "Store PaymentItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "PaymentItem that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/PaymentItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PaymentItem"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/paymentItems/{id}": {
            "get": {
                "tags": [
                    "PaymentItem"
                ],
                "summary": "Display the specified PaymentItem",
                "description": "Get PaymentItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PaymentItem",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PaymentItem"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "PaymentItem"
                ],
                "summary": "Update the specified PaymentItem in storage",
                "description": "Update PaymentItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PaymentItem",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "PaymentItem that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/PaymentItem"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/PaymentItem"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "PaymentItem"
                ],
                "summary": "Remove the specified PaymentItem from storage",
                "description": "Delete PaymentItem",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of PaymentItem",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "PaymentItem": {
            "required": [
                "payment_id",
                "amount"
            ],
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64"
                },
                "payment_id": {
                    "type": "integer"
                },
                "concept": {
                    "type": "string"
                },
                "amount": {
                    "type": "number",
                    "format": "double"
                }
            }
        },
...
    }
}